fathy-llm/requirements.txt
# Core deep learning runtime
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0

# Distributed + performance
accelerate>=0.33.0
deepspeed>=0.15.0
ninja>=1.11.1

# LLM ecosystem
transformers>=4.45.0
tokenizers>=0.20.0
datasets>=3.0.0
sentencepiece>=0.2.0
safetensors>=0.4.5
huggingface-hub>=0.24.0

# Alignment / RLHF
trl>=0.11.0
peft>=0.12.0

# Training utilities
einops>=0.8.0
scipy>=1.13.0
pyyaml>=6.0.2
tqdm>=4.66.0
wandb>=0.18.0
tensorboard>=2.17.0

# Quality / tests
pytest>=8.3.0
pytest-cov>=5.0.0

# Text quality / multilingual preprocessing
ftfy>=6.2.3
langdetect>=1.0.9
datasketch>=1.6.5
regex>=2024.9.11
arabic-reshaper>=3.0.0
python-bidi>=0.6.3

# Evaluation metrics
sacrebleu>=2.4.3
rouge-score>=0.1.2
